// exam/theia/ws-client/src/generated/openapi/model/answerFields.ts
/**
 * exam-back
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class AnswerFields {
    'question'?: boolean;
    'id'?: boolean;
    'content'?: boolean;
    'user'?: boolean;
    'judge'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "question",
            "baseName": "question",
            "type": "boolean"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "boolean"
        },
        {
            "name": "content",
            "baseName": "content",
            "type": "boolean"
        },
        {
            "name": "user",
            "baseName": "user",
            "type": "boolean"
        },
        {
            "name": "judge",
            "baseName": "judge",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return AnswerFields.attributeTypeMap;
    }
}