// exam/theia/ws-client/src/generated/openapi/model/newRecord.ts
/**
 * exam-back
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

/**
* (tsType: Record, schemaOptions: { title: \'NewRecord\' })
*/
export class NewRecord {
    'answer': number;
    'startTime': string;
    'endTime': string;
    'timeline': Array<string>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "answer",
            "baseName": "answer",
            "type": "number"
        },
        {
            "name": "startTime",
            "baseName": "startTime",
            "type": "string"
        },
        {
            "name": "endTime",
            "baseName": "endTime",
            "type": "string"
        },
        {
            "name": "timeline",
            "baseName": "timeline",
            "type": "Array<string>"
        }    ];

    static getAttributeTypeMap() {
        return NewRecord.attributeTypeMap;
    }
}