// exam/theia/ws-client/src/generated/openapi/model/questionFields.ts
/**
 * exam-back
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class QuestionFields {
    'id'?: boolean;
    'title'?: boolean;
    'author'?: boolean;
    'score'?: boolean;
    'brief'?: boolean;
    'detail'?: boolean;
    'examples'?: boolean;
    'examination'?: boolean;
    'templateName'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "boolean"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "boolean"
        },
        {
            "name": "author",
            "baseName": "author",
            "type": "boolean"
        },
        {
            "name": "score",
            "baseName": "score",
            "type": "boolean"
        },
        {
            "name": "brief",
            "baseName": "brief",
            "type": "boolean"
        },
        {
            "name": "detail",
            "baseName": "detail",
            "type": "boolean"
        },
        {
            "name": "examples",
            "baseName": "examples",
            "type": "boolean"
        },
        {
            "name": "examination",
            "baseName": "examination",
            "type": "boolean"
        },
        {
            "name": "templateName",
            "baseName": "templateName",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return QuestionFields.attributeTypeMap;
    }
}